requirements.txt
Django==4.2.10
djangorestframework==3.14.0
python-decouple==3.8
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
